토큰 방식 -> 쿠키 HttpUtil 활용 추가

+++ src/main/java/com/apiboad6/reactjwttokenproject/service/BoardService.java
@@ -1,14 +1,19 @@
 package com.apiboad6.reactjwttokenproject.service;
 
+import com.apiboad6.reactjwttokenproject.config.jwt.CustomUserDetails;
 import com.apiboad6.reactjwttokenproject.domain.board.Board;
 import com.apiboad6.reactjwttokenproject.domain.member.Member;
 import com.apiboad6.reactjwttokenproject.dto.board.BoardResponse;
 import com.apiboad6.reactjwttokenproject.dto.board.BoardUpdateRequest;
 import com.apiboad6.reactjwttokenproject.dto.board.BoardWriteRequest;
+import com.apiboad6.reactjwttokenproject.exception.ForbiddenActionException;
 import com.apiboad6.reactjwttokenproject.exception.NotFoundBoardException;
+import com.apiboad6.reactjwttokenproject.exception.NotLoginException;
 import com.apiboad6.reactjwttokenproject.repository.BoardRepository;
 import com.apiboad6.reactjwttokenproject.repository.SignRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -53,9 +58,24 @@ public class BoardService {
         board.update(req.getTitle(), req.getContent());
     }
 
+    //요청 하기전 미리 걸러주는 검증이기에 구분이 필요하다
     @Transactional
     public void deleteBoard(int id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        if(authentication == null || !authentication.isAuthenticated()) {
+            throw new NotLoginException();
+        }
+        CustomUserDetails user = (CustomUserDetails) authentication.getPrincipal();
+        int loginMemberId = user.getId();
+        String role = user.getRoleName();
+
         Board board = boardRepository.findById(id).orElseThrow(NotFoundBoardException::new);
+
+        if (loginMemberId != board.getMember().getId() && !role.equals("ROLE_ADMIN")) {
+            throw new ForbiddenActionException();
+        }
+
         boardRepository.delete(board);
     }
 }
